requirements-dev.txt
# Core Dashboard Framework
dash==2.14.2
dash-bootstrap-components==1.5.0
plotly==5.18.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database Connectors
snowflake-connector-python==3.6.0
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Environment & Configuration
python-dotenv==1.0.0

# Jupyter for Analysis
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Data Visualization
seaborn==0.13.0
matplotlib==3.8.2

# Utilities
openpyxl==3.1.2  # For Excel export
python-dateutil==2.8.2

# Optional: Progress bars and formatting
tqdm==4.66.1
tabulate==0.9.0
